--- src/__init__.py
@@ -1,14 +1,0 @@
-from config import (
-    VERSION,
-    CONSOLE,
-    DEBUG,
-    SOURCES,
-    WORDLIST,
-    SCRIPT,
-    HOST,
-    PORT,
-    WORD_MULT,
-    GEMS_MULT,
-    TRIE,
-    SWAP,
-)
