--- src/interfaces/webapi/__init__.py
@@ -1,4 +1,0 @@
-from apirouter import SolverRouter
-from baseapi import BaseAPI
-from baserouter import BaseRouter
-from interfaces import Response, SolverData
