--- src/modules/__init__.py
@@ -1,1 +1,0 @@
-from spellsolver import SpellSolver
