--- src/modules/gameboard/__init__.py
@@ -1,5 +1,0 @@
-from gameboard import GameBoard
-from gametile import GameTile
-from path import Path
-from resultlist import ResultList
-from resultword import ResultWord
