+++ src/modules/gameboard/gameboard.py
@@ -1,7 +1,7 @@
 from typing import Dict, Tuple
 
-from src.modules.gameboard import GameTile
-from src.utils import (
+from src.modules.gameboard.gametile import GameTile
+from src.utils.utils import (
     aux_to_indices,
     is_valid_word,
 )
