+++ src/modules/gameboard/path.py
@@ -1,9 +1,10 @@
 from typing import List, Tuple
 
-from src.modules.gameboard import GameTile
+from src.modules.gameboard.gameboard import GameTile
 
 
 class Path:
+
     @staticmethod
     def word_points(path: Tuple[GameTile, ...]) -> int:
         """Get points value of actual word"""
@@ -21,13 +22,9 @@ class Path:
         return sum(node.gems() for node in path)
 
     @staticmethod
-    def get_path(
-        path: Tuple[GameTile, ...], word: str, swaps: List[int]
-    ) -> Tuple[GameTile, ...]:
+    def get_path(path: Tuple[GameTile, ...], word: str, swaps: List[int]) -> Tuple[GameTile, ...]:
         """Get a new path with swap nodes replaced"""
         if not swaps:
             return path
 
-        return tuple(
-            (node.copy(word[i]) if i in swaps else node) for i, node in enumerate(path)
-        )
+        return tuple((node.copy(word[i]) if i in swaps else node) for i, node in enumerate(path))
