--- src/modules/trie/__init__.py
@@ -1,6 +1,0 @@
-from loader import swap_iter, word_iter, pair_iter
-from marisa import MarisaTrie, MarisaTrieQuery
-from patricia import PatriciaNode, PatriciaTrie, PatriciaTrieQuery
-from prefix import PrefixNode, PrefixTrie, PrefixTrieQuery
-from trie import Trie
-from trie_query import TrieQuery
