--- src/modules/validate/__init__.py
@@ -1,3 +1,0 @@
-from generate import WordGenerate
-from validate import WordValidate
-from wordlist import WordList
