--- src/utils/__init__.py
@@ -1,5 +1,0 @@
-from resources import resource_path
-from timer import Timer
-from timer_test import TestTimer
-from utils import CHARS, POINTS, aux_to_indices, is_valid_word, get_letter_point_value
-from utils_test import TestFunctions
