--- tests/__init__.py
@@ -1,2 +1,0 @@
-from test_graphicui import TestGraphicalUI
-from test_validate import Validate
